+++ renet2/src/channel/mod.rs
@@ -9,16 +9,16 @@ pub(crate) use slice_constructor::SliceConstructor;
 /// Delivery guarantee of a channel
 #[derive(Debug, Clone)]
 pub enum SendType {
-    // Messages can be lost or received out of order.
-    Unreliable,
+    /// Messages can be lost or received out of order.
+    Unreliable {
+        /// Should only be `true` if the underlying transport is ordered and reliable (e.g. WebSockets or in-memory channels)
+        /// and reliability is desired. Can be set to `false` if reliability is not required.`
+        ordered_reliable_substrate: bool,
+    },
     /// Messages are guaranteed to be received and in the same order they were sent.
-    ReliableOrdered {
-        resend_time: Duration,
-    },
+    ReliableOrdered { resend_time: Duration },
     /// Messages are guaranteed to be received but may be in an different order that they were sent.
-    ReliableUnordered {
-        resend_time: Duration,
-    },
+    ReliableUnordered { resend_time: Duration },
 }
 
 /// Configuration of a channel for a server or client
@@ -60,7 +60,9 @@ impl DefaultChannel {
             ChannelConfig {
                 channel_id: 0,
                 max_memory_usage_bytes: 5 * 1024 * 1024,
-                send_type: SendType::Unreliable,
+                send_type: SendType::Unreliable {
+                    ordered_reliable_substrate: false,
+                },
             },
             ChannelConfig {
                 channel_id: 1,